# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Moguchart REST API
  version: 1.1.0
  description: |
    Moguchartの外部アプリ向けREST API。
    ガントチャートプロジェクトの管理（プロジェクト・行・タスク・コメント・スナップショット）を提供します。

    ## 認証
    すべてのエンドポイント（ヘルスチェックを除く）は `X-API-Key` ヘッダーによるAPIキー認証が必要です。
    APIキーには `read`（読み取り専用）と `read-write`（読み書き）の2つのスコープがあります。
    書き込み操作には `read-write` スコープが必要です。

    ## レスポンス形式
    すべてのレスポンスは以下の統一形式で返されます:
    ```json
    {
      "status": "succeeded" | "failed",
      "message": "（任意）",
      "data": ...
    }
    ```
  contact:
    name: Moguchart
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api-{projectId}.cloudfunctions.net/api
    description: Firebase Cloud Functions (本番)
    variables:
      projectId:
        default: moguchart
        description: Firebase プロジェクトID
...
paths:
...
  /api/v1/tasks:
    post:
      operationId: createTasks
      summary: タスクの作成/更新
      description: タスクを作成または更新します。配列で複数タスクの一括処理が可能です。`read-write` スコープが必要です。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    tasks:
                      type: array
                      items:
                        $ref: '#/components/schemas/GanttTaskInput'
                  required: [tasks]
                - type: array
                  items:
                    $ref: '#/components/schemas/GanttTaskInput'
                - $ref: '#/components/schemas/GanttTaskInput'
      responses:
        '201':
          description: 作成/更新成功（タスクID配列を返す）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [succeeded]
                  data:
                    type: array
                    items:
                      type: integer
                    description: 作成/更新されたタスクID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      operationId: updateTasks
      summary: タスクの一括/単体更新
      description: タスクを更新します。配列で複数タスクの一括処理が可能です。`read-write` スコープが必要です。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  properties:
                    tasks:
                      type: array
                      items:
                        $ref: '#/components/schemas/GanttTaskInput'
                  required: [tasks]
                - type: array
                  items:
                    $ref: '#/components/schemas/GanttTaskInput'
                - $ref: '#/components/schemas/GanttTaskInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [succeeded]
                  data:
                    type: array
                    items:
                      type: integer
                    description: 更新されたタスクID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteTasks
      summary: タスクの一括削除
      description: 指定したタスクIDのタスクを一括削除します。`read-write` スコープが必要です。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                  description: 削除するタスクIDの配列
                  example: [10, 20]
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_Void'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    put:
      operationId: updateTaskById
      summary: タスクの単体更新
      description: 指定したタスクを更新します。`read-write` スコープが必要です。
      tags: [Tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GanttTaskInput'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [succeeded]
                  data:
                    type: array
                    items:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      operationId: deleteTaskById
      summary: タスクの単体削除
      description: 指定したタスクを削除します。`read-write` スコープが必要です。
      tags: [Tasks]
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_Void'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks/{taskId}/comments:
    parameters:
      - $ref: '#/components/parameters/TaskId'

    get:
      operationId: getTaskCommentsDirect
      summary: タスクコメント一覧取得
      description: 指定タスクに紐づくコメント一覧を取得します。
      tags: [Task Comments]
      responses:
        '200':
          description: タスクコメント一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_CommentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createTaskCommentDirect
      summary: タスクコメント作成/更新
      description: タスクに紐づくコメントを作成または更新します。`read-write` スコープが必要です。
      tags: [Task Comments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: 作成成功（コメントIDを返す）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [succeeded]
                  data:
                    type: integer
                    description: コメントID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks/{taskId}/comments/{commentId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
      - $ref: '#/components/parameters/CommentId'

    delete:
      operationId: deleteTaskCommentDirect
      summary: タスクコメント削除
      description: 指定したタスクコメントを削除します。`read-write` スコープが必要です。
      tags: [Task Comments]
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_Void'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ──────────────────────────────────────────────
  # Comments
  # ──────────────────────────────────────────────
...
  /api/v1/comments/tasks/{taskId}/comments:
    parameters:
      - name: taskId
        in: path
        required: true
        description: タスクID
        schema:
          type: integer

    get:
      operationId: listTaskComments
      summary: タスクコメント取得
      description: 指定タスクに紐づくコメント一覧を取得します。
      tags: [Task Comments]
      responses:
        '200':
          description: タスクコメント一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_CommentList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createTaskComment
      summary: タスクコメント作成/更新
      description: タスクに紐づくコメントを作成または更新します。`read-write` スコープが必要です。
      tags: [Task Comments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentInput'
      responses:
        '201':
          description: 作成成功（コメントIDを返す）
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [succeeded]
                  data:
                    type: integer
                    description: コメントID
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/comments/tasks/comments/{commentId}:
    parameters:
      - name: commentId
        in: path
        required: true
        description: タスクコメントID
        schema:
          type: integer

    delete:
      operationId: deleteTaskComment
      summary: タスクコメント削除
      description: タスクコメントを削除します。`read-write` スコープが必要です。
      tags: [Task Comments]
      responses:
        '200':
          description: 削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_Void'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ──────────────────────────────────────────────
  # Snapshots
  # ──────────────────────────────────────────────
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        `mk_` プレフィックス付きのAPIキー。
        スコープは `read`（読み取り専用）または `read-write`（読み書き）。
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Firebase IDトークン（JWT）。WebUIや管理操作向け。フルアクセス権限を持ちます。

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      description: タスクID
      schema:
        type: integer
        example: 10
    CommentId:
      name: commentId
      in: path
      required: true
      description: コメントID
      schema:
        type: integer
        example: 100
...
  responses:
    Unauthorized:
      description: 認証エラー — APIキーが未指定または無効
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: failed
            message: 'API key is required. Set the X-API-Key header.'
    Forbidden:
      description: 権限エラー — read-write スコープが必要
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: failed
            message: 'This API key has read-only access. A read-write key is required for this operation.'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [failed]
        message:
          type: string
    SuccessResponse_Void:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [succeeded]
...
    SuccessResponse_CommentList:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [succeeded]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    ColorPalette:
      type: object
      required: [color, backgroundColor]
      properties:
        name:
          type: string
          description: パレット名
        color:
          type: string
          description: テキスト色（CSS color）
          example: '#ffffff'
        backgroundColor:
          type: string
          description: 背景色（CSS color）
          example: '#42a5f5'
        pattern:
          type: object
          properties:
            type:
              type: string
              description: パターンの種類
            color:
              type: string
              description: パターンの色
        borderType:
          type: string
          enum: [solid_thin, solid_thick, dashed_thin, dashed_thick, dotted_thin, dotted_thick]
          description: 枠線の種類
        borderColor:
          type: string
          description: 枠線の色
    Label:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
          description: ラベル名
        color:
          type: string
          description: カラーコード
          example: '#ef5350'
...
    TaskAttribute:
      type: object
      description: タスクの追加属性
      properties:
        description:
          type: string
          description: タスクの説明
        colorPalette:
          $ref: '#/components/schemas/ColorPalette'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        lock:
          type: boolean
          description: ロック状態（移動・リサイズ・削除不可）
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: 進捗率（0〜100）
        dependencies:
          type: array
          items:
            type: string
          description: 依存タスクのID配列
        imageUrls:
          type: array
          items:
            type: string
          description: タスクバーに表示する画像のURL配列
...
    GanttTaskInput:
      type: object
      required: [rowId, name, start, end]
      properties:
        id:
          type: integer
          description: タスクID（更新時に指定）
        rowId:
          type: integer
          description: 所属する行のID
        name:
          type: string
          description: タスク名
        start:
          type: string
          format: date
          description: 開始日（YYYY-MM-DD）
        end:
          type: string
          format: date
          description: 終了日（YYYY-MM-DD）
        attribute:
          $ref: '#/components/schemas/TaskAttribute'
...
    Comment:
      type: object
      required: [id, content]
      properties:
        id:
          type: integer
          description: コメントID
        taskId:
          type: integer
          description: タスクID（タスクコメント時）
        rowId:
          type: integer
          description: 行ID（行コメント時）
        projectId:
          type: string
          description: プロジェクトID（プロジェクトコメント時）
        content:
          type: string
          description: コメント内容
        createdBy:
          type: string
          description: 作成者のメールアドレス
        createdByDisplayName:
          type: string
          description: 作成者の表示名
        createdByPhotoURL:
          type:
            - string
            - 'null'
          description: 作成者のアバター画像URL
        createdAt:
          type: string
          format: date-time
          description: 作成日時（ISO 8601）
    CommentInput:
      type: object
      required: [content]
      properties:
        id:
          type: integer
          description: コメントID（更新時に指定）
        taskId:
          type: integer
          description: タスクID
        rowId:
          type: integer
          description: 行ID
        projectId:
          type: string
          description: プロジェクトID
        content:
          type: string
          description: コメント内容
          example: '進捗確認お願いします'
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []